output: explicitly disable mouse and enable cursor on exit
SAVE_PRIVATE_MODE_VALUES and RESTORE_PRIVATE_MODE_VALUES when passed
with no modes are a Kitty extension to XTSAVE and XTRESTORE which
saves/restores *all* private modes. To my knowledge, Kitty is the only
terminal which implements this extension.CLEAR_SCREEN

As a result, when exiting awrit in a terminal which *does not* use this
extensions, the cursor is left hidden and mouse mode is left enabled.
Explicitly disable requested mouse reporting modes and show the cursor.

Reference: https://github.com/ghostty-org/ghostty/discussions/2732